refactor(analysis): extract error clustering into internal/analysis/clustering
internal/analysis was four unrelated subsystems in one bag. The api_*/
clustering_*/thirdparty_* filename prefixes are not independent evidence of
that — they were minted by one file-splitting commit (58264122 "split server
internals into focused modules"), so filenames agreeing with history is one
fact counted twice. The independent signal is the reference graph, and it
agrees: across all 23 non-test files there is not a single symbol reference
between the clustering files and the rest of the package. Co-change agrees
too — dfc42c29 (4 files) touched clustering and nothing else.

Moves the six clustering files to their own package, dropping the now-
redundant filename prefix, and takes their tests along: clustering_test.go,
clustering_unit_test.go (-> cluster_ops_test.go) and the TestMatchesCluster
block that was living in the package-wide branch_coverage_test.go.

Coverage of the new package is 96.9% with no function at 0.0%, so nothing
was left uncounted by the move.

internal/analysis: 28 -> 22 source files.
