+++ internal/analysis/clustering/normalize.go
@@ -1,6 +1,8 @@
 // Purpose: Normalizes error messages for stable cluster fingerprinting by replacing dynamic tokens.
 // Why: Separates message normalization regexes from cluster matching to keep both focused.
-package analysis
+// Docs: docs/features/feature/error-clustering/index.md
+
+package clustering
 
 import "regexp"
